microservices/agent-wizard/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
pyyaml==6.0.1
google-cloud-bigquery==3.11.4
google-auth==2.23.0

# LLM Integration
google-cloud-aiplatform>=1.38.0
openai>=1.3.0
anthropic>=0.7.0

# Additional utilities
aiohttp>=3.9.0
requests>=2.31.0